+++ lib/mongoid/attributes.rb
@@ -6,6 +6,7 @@ module Mongoid #:nodoc:
   # This module contains the logic for handling the internal attributes hash,
   # and how to get and set values.
   module Attributes
+    extend ActiveSupport::Concern
     include Processing
 
     attr_reader :attributes
@@ -97,11 +98,12 @@ module Mongoid #:nodoc:
     def write_attribute(name, value)
       assigning do
         access = name.to_s
+        localized = fields[access].try(:localized?)
         typed_value_for(access, value).tap do |value|
           unless attributes[access] == value || attribute_changed?(access)
             attribute_will_change!(access)
           end
-          attributes[access] = value
+          localized ? attributes[access].merge!(value) : attributes[access] = value
         end
       end
     end
@@ -200,6 +202,33 @@ module Mongoid #:nodoc:
     # @since 1.0.0
     def typed_value_for(key, value)
       fields.has_key?(key) ? fields[key].serialize(value) : value
+    end
+
+    module ClassMethods #:nodoc:
+
+      # Alias the provided name to the original field. This will provide an
+      # aliased getter, setter, existance check, and all dirty attribute
+      # methods.
+      #
+      # @example Alias the attribute.
+      #   class Product
+      #     include Mongoid::Document
+      #     field :price, :type => Float
+      #     alias_attribute :cost, :price
+      #   end
+      #
+      # @param [ Symbol ] name The new name.
+      # @param [ Symbol ] original The original name.
+      #
+      # @since 2.3.0
+      def alias_attribute(name, original)
+        class_eval <<-RUBY
+          alias :#{name} :#{original}
+          alias :#{name}= :#{original}=
+          alias :#{name}? :#{original}?
+        RUBY
+        super
+      end
     end
   end
 end