+++ spec/functional/mongoid/relations/synchronization_spec.rb
@@ -291,7 +291,7 @@ describe Mongoid::Relations::Synchronization do
     let!(:person) do
       Person.create(
         :ssn => "342-12-2222",
-        :preference_ids => [ one.id, two.id ]
+        :preferences => [ one, two ]
       )
     end