+++ spec/unit/mongoid/fields/serializable/integer_spec.rb
@@ -3,7 +3,7 @@ require "spec_helper"
 describe Mongoid::Fields::Serializable::Integer do
 
   let(:field) do
-    described_class.new(:test, :type => Integer)
+    described_class.instantiate(:test, :type => Integer)
   end
 
   describe "#deserialize" do
@@ -28,6 +28,20 @@ describe Mongoid::Fields::Serializable::Integer do
 
         it "returns the decimal" do
           field.serialize(2.5).should == 2.5
+        end
+      end
+
+      context "when the value is floating point zero" do
+
+        it "returns the integer zero" do
+          field.serialize(0.00000).should eq(0)
+        end
+      end
+
+      context "when the value is a floating point integer" do
+
+        it "returns the integer number" do
+          field.serialize(4.00000).should eq(4)
         end
       end
     end